+++ src/main/java/com/kamesuta/mc/autoinput/gui/InputKey.java
@@ -1,0 +1,5 @@
+package com.kamesuta.mc.autoinput.gui;
+
+public class InputKey {
+
+}
