+++ src/main/java/com/kamesuta/mc/autoinput/reference/Names.java
@@ -10,7 +10,6 @@ public class Names {
 		public static final String TITLE = "autoinputmod.gui.title";
 		public static final String SIWTCH_1 = "autoinputmod.gui.switch_hold";
 		public static final String SIWTCH_2 = "autoinputmod.gui.switch_continuous";
-		public static final String NONE = "autoinputmod.gui.none";
 		public static final String OPTIONS = "autoinputmod.gui.options";
 		public static final String DONE = "autoinputmod.gui.done";
 		public static final String OPTIONTEXT = "autoinputmod.gui.optiontext";
